FIX/cli/use.go: check if "bin" dir is existant before deleting it
I had this issue on my windows machine: zvm would never create the inital "bin"
directory (symlink) in ~/.zvm, as it always tries to remove the "bin" directory first.

Now, we're checking if the bin alias is even present and only delete it if it
exists.

Then afterwards the symlink to the selected version is created.