FIX/cli/meta/symlink_win.go: proper error check
On my Windows machine the call `errors.Is(err, &os.LinkError{})`, here:
https://github.com/tristanisham/zvm/blob/a7f1a2c429831b07d2b2dbd4eaf39ab212ec77cb/cli/meta/symlink_win.go#L57

always returned "false", so it was never creating the "bin" symlink at all.
Not a proficient Go dev here, but it looks like we need to compare the actuall
error to a syscall error and handle these?

Anyway I solved it differently by checking for `isAdmin()` first, then try to
become admin if not already and otherwise just create the link to "bin". This
way I even get the Windows User Account prompt if I want to allow the program to
make the change.