feat: add "run" command to invoke a compiler
this commit adds a "run" command to zvm that allows you to invoke a zig
compiler via its version name. for example: `zvm run 0.14.0 build run --help`
which is useful if you switch around projects a lot and just want to
run a single command without symlinking the version first.