fix flaky test on GitHub actions
This was caused by hitting rate limits, as GitHub Actions runners are pooled, they hit
limits quickly. This likely was effecting Mac and not the others due to the way Apple
requires physical hardware, so IP addresses are being reused all the time, unlike
Windows/Linux which are container based and can use unique IPs.

Three changes were made to address the issue:

1. The upgrade test was updated to fail immediately and obviously if a tag name is not
   returned
2. getLatestGitHubRelease() honors 403 and 429 failures, returning an error instead
   of silently failing
3. getLatestGitHubRelease() will use the GITHUB_TOKEN environment variable to acheive
   higher rate limits

Additionally, retry logic was put in place. This is likely unnecessary but there did not
seem to be a lot of reason to remove it. Also, some additional debug logs